+++ src/model/mapping/mapping_result_model.py
@@ -1,9 +1,13 @@
+import uuid
+
 from pandas import read_sql, DataFrame
 from sqlalchemy import text
 from sqlmodel import Session
 
+from repository.mapping.mapping_result_repository import save_mapping_results_list
 from repository.tenant.tenant_repository import get_tenant_by_id
 from scheme.mapping.mapping_results_scheme import MappingResult
+from scheme.mapping.mapping_scheme import MappingOneNomenclatureRead
 from scheme.user.user_scheme import UserRead
 
 
@@ -27,7 +31,21 @@ def _fetch_items(db_con_str: str, table_name: str, query: str) -> DataFrame:
     return read_sql(st, db_con_str)
 
 
-def get_mapping_result_by_id(mapping_result_id: int, session) -> MappingResult:
-    # TODO: raise exception when mapping_result is None
-    mapping_result = session.get(MappingResult, mapping_result_id)
-    return mapping_result
+def save_mapping_result(nomenclatures: list[MappingOneNomenclatureRead], user_id: int):
+    iteration_key = generate_iteration_key()
+
+    mapping_results = []
+    for nomenclature in nomenclatures:
+        mapping_result_dict = nomenclature.dict()
+        mapping_result = MappingResult(
+            user_id=user_id,
+            mapping_result=mapping_result_dict,
+            iteration_key=iteration_key,
+        )
+        mapping_results.append(mapping_result)
+
+    return save_mapping_results_list(mapping_results)
+
+
+def generate_iteration_key() -> str:
+    return str(uuid.uuid4())
