+++ src/scheme/mapping/mapping_results_scheme.py
@@ -18,6 +18,7 @@ class MappingResult(SQLModel, table=True):
     mapping_result: dict = Field(sa_column=Column(JSONB))
     created_at: datetime | None = Field(default_factory=datetime.utcnow)
     mapping_nomenclature_corrected: str | None = Field(default=None)
+    iteration_key: str | None = Field(default=None)
 
 
 class MappingResultUpdate(SQLModel):
